Make 64-bit int work with 'long int' and 'int64_t' for 64 bit builds (Issue #64)

+++ include/json/config.h
@@ -87,6 +87,7 @@
 #define JSONCPP_DEPRECATED(message)
 #endif // if !defined(JSONCPP_DEPRECATED)
 
+#include <stdint.h>
 namespace Json {
 typedef int Int;
 typedef unsigned int UInt;
@@ -100,8 +101,8 @@ typedef unsigned int LargestUInt;
 typedef __int64 Int64;
 typedef unsigned __int64 UInt64;
 #else                 // if defined(_MSC_VER) // Other platforms, use long long
-typedef long long int Int64;
-typedef unsigned long long int UInt64;
+typedef int64_t Int64;
+typedef uint64_t UInt64;
 #endif // if defined(_MSC_VER)
 typedef Int64 LargestInt;
 typedef UInt64 LargestUInt;
